+++ apps/wasm/task.py
@@ -20,165 +20,9 @@ from golem.task.taskbase import Task, AcceptClientVerdict
 from golem.task.taskstate import SubtaskStatus
 from golem.task.taskclient import TaskClient
 
+from .vbr import Actor, BucketVerifier, VerificationResult
+
 logger = logging.getLogger("apps.wasm")
-
-
-class Actor:
-    def __init__(self, uuid: str) -> None:
-        self.uuid = uuid
-
-
-class VerificationResult(IntEnum):
-    SUCCESS = 0
-    FAIL = 1
-    UNDECIDED = 2
-
-
-class VerificationByRedundancy(ABC):
-    def __init__(self, redundancy_factor: int, comparator: Callable[[Any, Any], int]) -> None:
-        self.redundancy_factor = redundancy_factor
-        assert comparator.func_closure is None
-        self.comparator = comparator
-
-    @abstractmethod
-    def add_actor(self, actor: Actor) -> None:
-        """Caller informs class that this is the next actor he wants to assign to the next subtask.
-
-        Raises:
-            NotAllowedError -- Actor given by caller is not allowed to compute next task.
-            MissingResultsError -- Raised when caller wants to add next actor but has already
-            exhausted this method. Now the caller should provide results by `add_result` method.
-        """
-
-    @abstractmethod
-    def add_result(self, actor: Actor, result: Optional[Any]) -> None:
-        """Add a result for verification.
-
-        If a task computation has failed for some reason then the caller should use this method with the
-        result equal to None.
-
-        When user has added a result for each actor it reported by `add_actor` a side effect might be
-        the verdict being available or caller should continue adding actors and results.
-
-        Arguments:
-            actor {Actor} -- Actor who has computed the result
-            result {Any} --  Computation result
-        Raises:
-            UnknownActorError - raised when caller deliver an actor that was not previously reported by `add_actor` call.
-        """
-
-    @abstractmethod
-    def get_verdicts(self) -> Optional[List[Tuple[Any, Actor, VerificationResult]]]:
-        """
-        Returns:
-            Optional[List[Any, Actor, VerificationResult]] -- If verification is resolved a list of 3-element
-            tuples (result reference, actor, verification_result) is returned. A None is returned
-            when verification has not been finished yet.
-        """
-        pass
-
-
-class SimpleSubtaskVerifier(VerificationByRedundancy):
-    """Simple verification by redundancy: subtask is executed by 2 providers,
-with a possible third if no decision can be reached based on the first 2."""
-
-    def __init__(self,
-                 redundancy_factor: int, comparator: Callable[[Any, Any], int]) -> None:
-        super().__init__(redundancy_factor, comparator)
-        self.actors = []
-        self.results = {}
-
-    def add_actor(self, actor):
-        """Caller informs class that this is the next actor he wants to assign to the next subtask.
-        Raises:
-            NotAllowedError -- Actor given by caller is not allowed to compute next task.
-            MissingResultsError -- Raised when caller wants to add next actor but has already
-            exhausted this method. Now the caller should provide results by `add_result` method.
-        """
-
-        if actor in self.actors:
-            raise NotAllowedError
-
-        # TODO: when should we raise MissingResultsError?
-
-    def add_result(self, actor: Actor, result: Optional[Any]) -> None:
-        """Add a result for verification.
-        If a task computation has failed for some reason then the caller should use this method with the
-        result equal to None.
-        When user has added a result for each actor it reported by `add_actor` a side effect might be
-        the verdict being available or caller should continue adding actors and results.
-        Arguments:
-            actor {Actor} -- Actor who has computed the result
-            result {Any} --  Computation result
-        Raises:
-            UnknownActorError - raised when caller deliver an actor that was not previously reported by `add_actor` call.
-            ValueError - raised when attempting to add a result for some actor more than once
-        """
-
-        if actor not in self.actors:
-            raise UnknownActorError
-
-        if actor in self.results:
-            raise ValueError
-
-        self.results[actor] = result
-
-    def get_verdicts(self) -> Optional[List[Tuple[Any, Actor, VerificationResult]]]:
-        actor_cnt = len(self.actors)
-        result_cnt = len(self.results.keys())
-
-        # get_verdicts should only get called when all results have been added
-        # calling it otherwise is a contract violation
-        assert(actor_cnt == result_cnt)
-
-        assert(actor_cnt <= 3)
-
-        if actor_cnt < 2:
-            return None
-
-        verdict_undecided = [
-            (a, self.results[a], VerificationResult.UNDECIDED) for a in self.actors]
-        real_results = [(a, self.results[a])
-                        for a in self.actors if self.results[a] is not None]
-        reporting_actors = [r[0] for r in real_results]
-        reported_results = [r[1] for r in real_results]
-
-        if actor_cnt == 2:
-            if len(real_results) == 0:
-                return verdict_undecided  # more actors won't help, giving up
-
-            if len(real_results) == 1:
-                return None  # not enough real results, need more actors
-
-        if actor_cnt > 2 and len(real_results) < 2:
-            return verdict_undecided
-
-        if len(real_results) == 2:
-            a1, a2 = reporting_actors
-            r1, r2 = reported_results
-
-            if self.comparator(r1, r2) == 0:
-                return [(a1, r1, VerificationResult.SUCCESS), (a2, r2, VerificationResult.SUCCESS)]
-            else:
-                if actor_cnt > 2:
-                    return verdict_undecided  # give up
-                else:
-                    return None  # more actors needed
-        else:  # 3 real results
-            a1, a2, a3 = reporting_actors
-            r1, r2, r3 = real_results
-
-            if self.comparator(r1, r2) == 0 and self.comparator(r2, r3) == 0 and self.comparator(r3, r1) == 0:
-                return [(r[0], r[1], VerificationResult.SUCCESS) for r in real_results]
-
-            for permutation in [(0, 1, 2), (1, 2, 0), (2, 0, 1)]:
-                permuted_pairs = [real_results[i] for i in permutation]
-                a1, a2, a3 = [p[0] for p in permuted_pairs]
-                r1, r2, r3 = [p[1] for p in permuted_pairs]
-                if self.comparator(r1, r2) == 0:  # r1 and r2 are equal, hence r3 differs
-                    return[(a1, r1, VerificationResult.SUCCESS), (a2, r2, VerificationResult.SUCCESS), (a3, r3, VerificationResult.FAIL)]
-
-            return verdict_undecided  # all 3 results different
 
 
 class VbrSubtask:
@@ -188,7 +32,7 @@ class VbrSubtask:
         self.params = params
 
         self.subtasks = {}
-        self.verifier = SimpleSubtaskVerifier(
+        self.verifier = BucketVerifier(
             redundancy_factor, WasmTask._cmp_results)
 
     def contains(self, s_id):
